Update reference.md
I've edited the doc for readability, flow, etc.

There are still a few things that need to be fixed. Some of the links don't work, or point to the wrong place. For example, "You can view a sample applications.json here" isn't a link at all, and should point to the correct sample file in the tutorial space. 

I know Dave already expressed issues with the debugging section. I wonder if we should be more prescriptive. Rather than saying, "The simplest solution to this is to add an JavaScript alert() followed by a Javascript debugger; statement." Perhaps we should give them the code and where exactly to put it.